+++ component-library/src/rich-text-editor/plugins/index.tsx
@@ -50,3 +50,6 @@ export const customPlugins: SlatePlugin[] = [
 ]
 
 export const basePlugins: SlatePlugin[] = [...genericPlugins, ...customPlugins]
+export * from './peak-callout-plugin/defaults'
+export * from './peak-divider/defaults'
+export * from './peak-media-embed-plugin/types'
