+++ component-library/src/rich-text-editor/plugins/peak-divider/defaults.ts
@@ -1,2 +1,1 @@
 export const ELEMENT_DIVIDER = 'divider'
-
