+++ peak-app/src/App.tsx
@@ -4,6 +4,7 @@ import { Provider } from "react-redux";
 import {PersistGate} from "redux-persist/integration/react";
 import {store, persistor} from "./redux/store";
 import ProvidedApp from "./ProvidedApp";
+import 'component-library/dist/index.css'
 
 function App() {
   return (
