+++ peak-app/src/common/peak-toolbar/toolbar-controls.tsx
@@ -35,7 +35,8 @@ import {
     UnderlineOutlined,
     UnorderedListOutlined,
     BookOutlined,
-    BulbOutlined
+    BulbOutlined,
+    LineOutlined
 } from "@ant-design/icons/lib";
 import { Icon, InlineIcon } from '@iconify/react';
 import headingH1 from '@iconify/icons-gridicons/heading-h1';
@@ -44,14 +45,13 @@ import headingH3 from '@iconify/icons-gridicons/heading-h3';
 import headingH4 from '@iconify/icons-gridicons/heading-h4';
 import headingH5 from '@iconify/icons-gridicons/heading-h5';
 import {message} from "antd";
-import {PEAK_CALLOUT} from "../rich-text-editor/plugins/peak-callout-plugin/defaults";
 import {insertCustomBlockElement} from "../rich-text-editor/utils/base-utils";
 import {PeakNodeSelectListItem} from "../rich-text-editor/utils/node-content-select/types";
 import {convertEditorControlDisplayToNodeSelectListItem} from "../rich-text-editor/utils/node-content-select/utils";
 import {PEAK_LEARNING, PEAK_BOOK_SELECT_ITEM} from "../rich-text-editor/plugins/peak-knowledge-plugin/constants";
 import {defaultOptions} from "../rich-text-editor/options";
-import {PEAK_DIVIDER} from "../rich-text-editor/plugins/peak-divider/constants";
 import {PEAK_MEDIA_EMBED_CONTROLS} from "../rich-text-editor/plugins/peak-media-embed-plugin/constants";
+import {ELEMENT_DIVIDER, PEAK_CALLOUT} from "component-library/dist";
 
 export interface PeakEditorControl {
     controlType: "mark" | "block" | "list" | "img" | "code_block" | undefined
@@ -147,6 +147,14 @@ const PEAK_LINK: PeakEditorControlDisplay = {
     elementType: ELEMENT_LINK,
     hotkeyInstructionArray: ['⌘', 'K'],
     customFormat: (editor => toggleMark(editor, ELEMENT_LINK)),
+};
+
+export const PEAK_DIVIDER: PeakEditorControlDisplay = {
+    controlType: "block",
+    icon: <LineOutlined className={"peak-editor-control-icon"}/>,
+    description: "Separate content with a horizontal line",
+    label: "Divider",
+    elementType: ELEMENT_DIVIDER,
 };
 // const PEAK_CODE_BLOCK: PeakEditorControlDisplay = {
 //     controlType: ELEMENT_CODE_BLOCK,
