+++ peak-app/src/common/rich-text-editor/components.tsx
@@ -19,7 +19,6 @@ import {
 } from "@udecode/slate-plugins";
 import {DragIndicator} from "@styled-icons/material/DragIndicator";
 import React, {useMemo} from "react";
-import {PEAK_CALLOUT} from "./plugins/peak-callout-plugin/defaults";
 import {PEAK_LEARNING} from "./plugins/peak-knowledge-plugin/constants";
 import {PeakLearningElement} from "./plugins/peak-knowledge-plugin/components/peak-learning-node/PeakLearningNode";
 import {PeakInlineLinkElement} from "./plugins/peak-link-plugin/inline-link/PeakHyperLink";
@@ -29,24 +28,7 @@ import {
 import {PeakMediaStubElement} from "./plugins/peak-media-embed-plugin/components/stub/MediaEmbedStub";
 import {clone} from "ramda";
 import {DRAGGABLE_ELEMENTS} from "./constants";
-import {PEAK_SLATE_COMPONENT_OVERRIDES} from "component-library/dist";
-
-// const defaultComponents = createSlatePluginsComponents({
-//     [ELEMENT_BLOCKQUOTE]: PeakBlockquoteElement,
-//     [ELEMENT_LI]: PEAK_LI_STYLE,
-//     [ELEMENT_UL]: PEAK_UL_STYLE,
-//     [ELEMENT_OL]: PEAK_OL_STYLE,
-//     [ELEMENT_CODE_BLOCK]: PeakCodeBlockElement,
-//     [TITLE]: PeakTitleElement,
-//     [PEAK_CALLOUT]: PeakCalloutElement,
-//     [PEAK_LEARNING]: PeakLearningElement,
-//     [ELEMENT_LINK]: PeakInlineLinkElement,
-//     [ELEMENT_EMBED_STUB]: PeakMediaStubElement,
-//     [ELEMENT_TWITTER_EMBED]: TwitterEmbed,
-//     [ELEMENT_YOUTUBE_EMBED]: YoutubeEmbed,
-//     [ELEMENT_MEDIA_EMBED]: RichLinkEmbed,
-//     [ELEMENT_DIVIDER]: DividerElement,
-// })
+import {PEAK_CALLOUT, PEAK_SLATE_COMPONENT_OVERRIDES} from "component-library/dist";
 
 const defaultComponents = createSlatePluginsComponents({
     ...PEAK_SLATE_COMPONENT_OVERRIDES,
