+++ peak-app/src/common/rich-text-editor/options.ts
@@ -5,7 +5,6 @@ import { ELEMENT_BLOCKQUOTE } from '@udecode/slate-plugins-block-quote';
 import { ELEMENT_CODE_BLOCK } from '@udecode/slate-plugins-code-block';
 import { SlatePluginOptions } from '@udecode/slate-plugins-core';
 import { ELEMENT_PARAGRAPH } from '@udecode/slate-plugins-paragraph';
-import {PEAK_CALLOUT} from "./plugins/peak-callout-plugin/defaults";
 import {ELEMENT_PEAK_BOOK, PEAK_LEARNING} from "./plugins/peak-knowledge-plugin/constants";
 import {JOURNAL_ENTRY, TITLE} from "./types";
 import {
@@ -15,7 +14,7 @@ import {
     SoftBreakPluginOptions,
 } from "@udecode/slate-plugins";
 import {options} from "../../views/scratchpad/playground/defaultOptions";
-import {ELEMENT_DIVIDER} from "./plugins/peak-divider";
+import {PEAK_CALLOUT, ELEMENT_DIVIDER } from "component-library/dist";
 
 const PEAK_STRIKETHROUGH_OPTIONS: Partial<SlatePluginOptions> = {
     hotkey: 'mod+shift+x',
