--- peak-app/src/common/rich-text-editor/plugins/peak-callout-plugin/PeakCalloutPlugin.ts
@@ -1,10 +1,0 @@
-import {
-    getRenderElement,
-    SlatePlugin,
-} from "@udecode/slate-plugins";
-import {PEAK_CALLOUT} from "./defaults";
-
-export const createPeakCalloutPlugin = (): SlatePlugin => ({
-    pluginKeys: PEAK_CALLOUT,
-    renderElement: getRenderElement(PEAK_CALLOUT),
-});
