--- peak-app/src/common/rich-text-editor/plugins/peak-callout-plugin/component/PeakCallout.tsx
@@ -1,9 +1,0 @@
-import {RenderElementProps} from "slate-react";
-import React from "react";
-import "./peak-callout.scss"
-
-export const PeakCalloutElement = (props: RenderElementProps) => {
-    return (
-        <div className={"peak-callout-container"} {...props.attributes}>{props.children}</div>
-    )
-}
