--- peak-app/src/common/rich-text-editor/plugins/peak-callout-plugin/component/peak-callout.scss
@@ -1,8 +1,0 @@
-.peak-callout-container {
-  background-color: rgba(235, 236, 237, 0.4);
-
-  display: flex;
-  border-radius: 3px;
-  padding: 12px 16px;
-  overflow: auto;
-}
