--- peak-app/src/common/rich-text-editor/plugins/peak-callout-plugin/defaults.ts
@@ -1,2 +1,0 @@
-export const PEAK_CALLOUT = "callout";
-export declare type PeakCalloutKeyOption = typeof PEAK_CALLOUT
