--- peak-app/src/common/rich-text-editor/plugins/peak-divider/DividerElement.types.ts
@@ -1,6 +1,0 @@
-import { RootStyleSet } from "@udecode/slate-plugins";
-import { IStyle } from '@uifabric/styling';
-
-export interface DividerElementStyleSet extends RootStyleSet {
-    hr?: IStyle
-}
