--- peak-app/src/common/rich-text-editor/plugins/peak-divider/createDividerPlugin.ts
@@ -1,7 +1,0 @@
-import {getRenderElement, SlatePlugin } from "@udecode/slate-plugins";
-import {ELEMENT_DIVIDER} from "./defaults";
-
-export const createDividerPlugin = (): SlatePlugin => ({
-    pluginKeys: ELEMENT_DIVIDER,
-    renderElement: getRenderElement(ELEMENT_DIVIDER),
-});
