--- peak-app/src/common/rich-text-editor/plugins/peak-divider/defaults.ts
@@ -1,2 +1,0 @@
-export const ELEMENT_DIVIDER = 'divider';
-
