--- peak-app/src/common/rich-text-editor/plugins/peak-divider/index.ts
@@ -1,1 +1,0 @@
-export * from './defaults';
