--- peak-app/src/common/rich-text-editor/plugins/peak-title-plugin/PeakTitlePlugin.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import {
-    getRenderElement,
-    SlatePlugin
-} from "@udecode/slate-plugins";
-import {TITLE} from "../../types";
-
-export const createPeakTitlePlugin = (): SlatePlugin => ({
-    pluginKeys: TITLE,
-    renderElement: getRenderElement(TITLE),
-});
