--- peak-app/src/common/rich-text-editor/plugins/peak-title-plugin/peak-title/peak-title.scss
@@ -1,15 +1,0 @@
-@import "../../../../../constants/variables";
-
-.page-title {
-  position: relative;
-  margin-bottom: 30px;
-  width: 100%;
-
-  &.empty-title::after {
-    position: absolute;
-    color: #8993a4;
-    top: 0;
-    cursor: text;
-    content: "Page Title";
-  }
-}
