+++ peak-app/src/common/rich-text-editor/types.ts
@@ -4,9 +4,9 @@ import {
     SPEditor,
 } from "@udecode/slate-plugins";
 import {PeakKnowledgeKeyOption} from "./plugins/peak-knowledge-plugin/types";
-import {PeakCalloutKeyOption} from "./plugins/peak-callout-plugin/defaults";
 import {ReactEditor} from "slate-react";
 import {HistoryEditor} from "slate-history";
+import { PeakCalloutKeyOption } from "component-library";
 
 export const JOURNAL_ENTRY = "journal_entry";
 export const JOURNAL_ENTRY_HEADER = "journal_entry_header";
